requirements.txt
appdirs==1.4.4
argcomplete==2.0.0
bidict==0.22.0
click==8.1.3
colorama==0.4.4
Flask==2.1.2
Flask-Cors==3.0.10
Flask-SocketIO==4.3.1
importlib-metadata==4.11.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
pycparser==2.21
pyserial==3.5
python-engineio==3.13.2
python-escpos==2.2.0
python-socketio==4.6.0
pyusb==1.2.1
PyYAML==6.0
qrcode==7.3.1
six==1.16.0
typing_extensions==4.2.0
Werkzeug<2.1
zipp==3.8.0

